refactor(hu-reviewer): extract splitting detection into phase module
Audit follow-up (KJC-PCS-0033, third HIGH god-function).

runHuReviewerStage was 311 LOC, doing too many things:
  load-stories → split-detection → certify-loop → topo-sort

The 78-LOC splitting block (lines 242-318) had its own non-trivial
control flow (per-story heuristic loop + autonomous vs interactive
modes + LLM proposal retries) and only touched the batch object —
classic extraction candidate.

Moved to:
  src/orchestrator/stages/hu-reviewer-phases/detect-and-apply-splits.js

Helpers introduced inside the phase module to remove duplication
between the autonomous and interactive paths:
  - buildSubHuStory(sub)   ─ canonical story shape from a proposal
  - applySplit({...})      ─ batch.splice + emit split-accepted event

Behaviour identical — same heuristic order, same autonomous-vs-FDE
branching, same persistence point, same emitted events.

LOC after extraction:
  hu-reviewer-stage.js     397 → 394 (function: 311 → 239 LOC, -23%)
  detect-and-apply-splits  +130 LOC (new file, includes JSDoc header)

Also drops two now-unused imports from hu-reviewer-stage.js
(detectIndicators/selectHeuristic, generateSplitProposal/format/
buildSplitDependencies) — they all moved with the logic.

Tests: 4037/4037 green. Dynamic-imports budget unchanged (≤150).
